% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{both_indexes}
\alias{both_indexes}
\title{Gather RCF AND OCF indexes in one data.frame}
\usage{
both_indexes(x, y)
}
\arguments{
\item{x}{data.frame RCF index table}

\item{y}{data.frame OCF index table}
}
\value{
data.frame with both indexes ordered according to common atributes
}
\description{
this function binds RCF and OCF indexes in two consecutives columns, in order to look at glance both indexes.
}
\author{
Omar Benites
}
