% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{ccf}
\alias{ccf}
\title{Community Cultivar Frequency (CCF)}
\usage{
ccf(dfr, vname, hh, community, location, pctn = TRUE)
}
\arguments{
\item{dfr}{data.frame table name that constains the data}

\item{vname}{numeric total number of household}

\item{hh}{character household column name}

\item{community}{character community name where. A community o communities belongs to certain locaiton}

\item{location}{character location or admin level column name where cultivars were sampled or monitored}

\item{pctn}{logical by default TRUE. FALSE if do not requiere in percentage mode}
}
\value{
it returns the RCF index and RCF scale
}
\description{
calculation of the frequency of varieties by community
}
\examples{
library(askha)
data(samdata)
dfr_hfc <- hcf(samdata, "code_farmer", "number_tuber_by_sampling")
}
\author{
Omar Benites
}
