% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcf.R
\name{hcf}
\alias{hcf}
\title{Household Cultivar Frequency (HCF)}
\usage{
hcf(dfr, hh, nsvarie, ncut, pctn = TRUE)
}
\arguments{
\item{dfr}{data.frame table name that constains the data}

\item{hh}{character household colum name}

\item{nsvarie}{numeric number of sampled varieties}

\item{ncut}{integer minimum (from the total) number of sampled varied}

\item{pctn}{logical by default TRUE. FALSE if do not requiere in percentage mode}
}
\value{
it returns the HCF
}
\description{
calculation of the frequency of cultivars by houesholds
}
\examples{
library(askha)
data(samdata)
dfr_hfc <- hcf(samdata, "code_farmer", "number_tuber_by_sampling")
}
\author{
Omar Benites
}
