% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf.R
\name{rcf}
\alias{rcf}
\title{Relative Cultivar Frequency (RCF)}
\usage{
rcf(dfr, vname, hh, nsvarie, location)
}
\arguments{
\item{dfr}{data.frame table name that constains the data}

\item{vname}{numeric total number of household}

\item{hh}{character household column name}

\item{nsvarie}{character number of sampled varieties}

\item{location}{character location or admin level column name where cultivars were sampled or monitored}
}
\value{
it returns the RCF index and RCF scale
}
\description{
calculation of the relative cultivar frequency by household
}
\examples{
library(askha)
data(samdata)
dfr_hfc <- hcf(samdata, "code_farmer", "number_tuber_by_sampling")
}
\author{
Omar Benites
}
