% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samdata}
\alias{samdata}
\title{Sample data to demostrate the calculation of RCF and OCF indeces}
\format{
A data frame with 5 columns and 5030 rows.
}
\source{
International Potato Center, potato data.
}
\usage{
samdata
}
\description{
This data set has  numeration of the observations (\code{numeration}),
variety name (\code{variety}),the code or name of households (\code{code_farmer}),
and the location or admin level (\code{ADM3_Name}), comunities of the locations (\code{cu_community}), the number of sampling tubers in households (\code{number_tuber_by_sampling})
}
