% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocf.R
\name{ocf}
\alias{ocf}
\title{Relative Cultivar Frequency (RCF)}
\usage{
ocf(dfr, vname, hh, community, location)
}
\arguments{
\item{dfr}{data.frame table name that constains the data}

\item{vname}{numeric total number of household}

\item{hh}{character household column name}

\item{community}{character community name where. A community o communities belongs to certain locaiton}

\item{location}{character location or admin level column name where cultivars were sampled or monitored}
}
\value{
it returns the RCF index and RCF scale
}
\description{
calculation of the overall cultivar frequency by household
}
\examples{
library(askha)
data(samdata)
dfr_hfc <- hcf(samdata, "code_farmer", "number_tuber_by_sampling")
}
\author{
Omar Benites
}
