% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf.R
\name{rcf}
\alias{rcf}
\title{Relative Cultivar Frequency (RCF)}
\usage{
rcf(
  dfr,
  vname,
  hh,
  nsvarie,
  community,
  location,
  is_grouped = FALSE,
  shorten = FALSE
)
}
\arguments{
\item{dfr}{data.frame table name that constains the data}

\item{vname}{numeric total number of household}

\item{hh}{character household column name}

\item{nsvarie}{character number of sampled varieties}

\item{community}{character vector of one or multiple communities that belong to locations}

\item{location}{character location or admin level column name where cultivars
were sampled or monitored}

\item{is_grouped}{logical whether group the number of varieties by location. By default \code{is_grouped=FALSE}}

\item{shorten}{logical \code{shorten=TRUE} only show a simplified version of the original data. Whether \code{shorten=TRUE} append OCF and OCF scales to original data}
}
\value{
it returns the RCF index and RCF scale
}
\description{
calculation of the relative cultivar frequency by household
}
\examples{
library(askha)
data(samdata)
rcf(samdata, "variety" ,"code_farmer", "number_tuber_by_sampling", "cu_community"  ,"ADM3_Name")
}
\author{
Omar Benites
}
